backend/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.7
django-cors-headers==3.13.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.0
PyJWT==1.7.1
pytz==2022.7.1
requests==2.28.2
ruamel.yaml==0.17.21
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
wincertstore==0.2
